# cultural-tourism-backend/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + Gin + 腾讯云开发构建的 RESTful API",
        "title": "数字文旅后端 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/favorites": {
            "get": {
                "description": "分页获取当前用户的收藏列表 (支持按资源类型筛选)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "获取用户收藏列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型筛选 (theme/poi/product)",
                        "name": "resource_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码 (默认1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量 (默认20, 最大100)",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "收藏列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "用户收藏旅拍主题/景点POI/商品 (幂等：重复收藏返回错误)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "收藏资源",
                "parameters": [
                    {
                        "description": "收藏请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FavoriteCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "收藏成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误或已收藏",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/comments": {
            "get": {
                "description": "默认只显示审核通过(status=1)的评论",
                "tags": [
                    "Comments"
                ],
                "summary": "获取评论列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "点位ID",
                        "name": "poi_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态 (1:通过, 0:待审)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "用户发布评论 (默认待审核 status=0)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "发布评论",
                "parameters": [
                    {
                        "description": "评论信息",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "获取评论详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "用于管理员审核 (修改status) 或 用户点赞 (修改like_count)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "更新评论 (审核/点赞)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容 (仅status/like_count)",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "管理员可删除违规评论，用户可删除自己的评论(依赖TCB权限)",
                "tags": [
                    "Comments"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/favorites/{resource_type}/{resource_id}": {
            "get": {
                "description": "检查指定资源是否已被当前用户收藏 (RESTful路径参数)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "检查是否已收藏",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型 (theme/poi/product)",
                        "name": "resource_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "资源ID",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回 {is_favorited: true/false}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "通过资源类型和资源ID取消收藏 (RESTful路径参数)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "取消收藏",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型 (theme/poi/product)",
                        "name": "resource_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "资源ID",
                        "name": "resource_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "收藏记录不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/photos": {
            "get": {
                "description": "支持按主题ID筛选，默认只显示审核通过(status=1)的照片",
                "tags": [
                    "Photos"
                ],
                "summary": "获取照片列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主题ID",
                        "name": "theme_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态 (1:通过, 0:待审)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "用户上传旅拍照片 (上传后默认为待审核状态 status=0)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos"
                ],
                "summary": "上传照片",
                "parameters": [
                    {
                        "description": "照片信息",
                        "name": "photo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Photo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/photos/{id}": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "获取照片详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "照片ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "用于管理员审核 (修改status) 或 用户点赞 (修改like_count)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Photos"
                ],
                "summary": "更新照片 (审核/点赞)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "照片ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容 (仅status/like_count)",
                        "name": "photo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Photo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "管理员可删除违规照片，用户可删除自己的照片(依赖TCB权限)",
                "tags": [
                    "Photos"
                ],
                "summary": "删除照片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "照片ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pois": {
            "get": {
                "description": "查询点位列表，支持按区域、类型筛选。若传入 lat/lng，结果将包含距离信息(_distance)。",
                "tags": [
                    "POI"
                ],
                "summary": "获取点位列表 (支持LBS)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "区域ID",
                        "name": "region_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "点位类型 (scenic/food/hotel/booth)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "用户纬度 (用于计算距离)",
                        "name": "lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "用户经度 (用于计算距离)",
                        "name": "lng",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的 POI 点位 (景点/饭店/酒店/旅拍机)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POI"
                ],
                "summary": "创建点位",
                "parameters": [
                    {
                        "description": "POI信息",
                        "name": "poi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.POI"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pois/{id}": {
            "get": {
                "tags": [
                    "POI"
                ],
                "summary": "获取点位详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.POI"
                        }
                    }
                }
            },
            "put": {
                "description": "更新点位信息 (安全更新，忽略系统字段)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POI"
                ],
                "summary": "更新点位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "poi",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.POI"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "POI"
                ],
                "summary": "删除点位",
                "parameters": [
                    {
                        "type": "string",
                        "description": "POI ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "商品导流列表（无支付，点击跳转）",
                "tags": [
                    "Products"
                ],
                "summary": "获取商品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建商品导流信息 (无支付，仅跳转)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "创建商品导流",
                "parameters": [
                    {
                        "description": "商品信息",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "获取商品详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新商品导流信息 (支持增量更新)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "更新商品导流",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "删除商品导流",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Comment": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "TCB 自动 ID",
                    "type": "string"
                },
                "_openid": {
                    "description": "发布者 (系统字段)",
                    "type": "string"
                },
                "content": {
                    "description": "评论内容",
                    "type": "string"
                },
                "created_at": {
                    "description": "业务创建时间",
                    "type": "string"
                },
                "like_count": {
                    "description": "点赞数",
                    "type": "integer"
                },
                "parent_id": {
                    "description": "父评论 ID (空字符串表示一级评论)",
                    "type": "string"
                },
                "poi_id": {
                    "description": "关联的点位 ID",
                    "type": "string"
                },
                "status": {
                    "description": "0:待审, 1:通过, 2:拒绝",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "业务更新时间",
                    "type": "string"
                }
            }
        },
        "models.FavoriteCreateRequest": {
            "type": "object",
            "required": [
                "resource_id",
                "resource_type"
            ],
            "properties": {
                "resource_id": {
                    "type": "string"
                },
                "resource_type": {
                    "description": "枚举验证",
                    "type": "string",
                    "enum": [
                        "theme",
                        "poi",
                        "product"
                    ]
                }
            }
        },
...
    }
}